Alter VideoInfoProvider to not create thumbnails or snapshots if a thumbnail already exists

+++ Engine/Providers/VideoInfoProvider.cs
@@ -1,5 +1,6 @@
 using System.Diagnostics;
 using System.IO.Abstractions;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Engine.Foundation;
@@ -30,16 +31,34 @@ namespace Engine.Providers
 
             var result = await CreateSnapshot(filePath, 30, token);
 
-            await this.thumbnailCreator.CreateAsync(result.SnapshotPath, result.SnapshotPath.Replace("_snap", string.Empty), token);
+            string thumbnailPath;
+            if (result.Created)
+            {
+                var thumbnailResult = await this.thumbnailCreator.CreateAsync(result.SnapshotPath,
+                    result.SnapshotPath.Replace("_snap", string.Empty), token);
 
-            this.fileSystem.File.Delete(result.SnapshotPath);
+                this.fileSystem.File.Delete(result.SnapshotPath);
+                thumbnailPath = thumbnailResult.OutputPath;
+            }
+            else
+            {
+                thumbnailPath = result.SnapshotPath;
+            }
 
-            return new VideoMetaData(filePath, result.SnapshotPath, 0, 0, info.Size, info.Duration);
+            var videoStream = info.VideoStreams.First();
+
+            return new VideoMetaData(filePath, thumbnailPath, videoStream.Width, videoStream.Height, info.Size, info.Duration);
         }
 
         public async Task<SnapshotResult> CreateSnapshot(string filePath, int seconds, CancellationToken token)
         {
             var cacheLocation = this.thumbnailCacheLocationProvider.ProvideLocation(filePath);
+
+            var checkPath = $"{cacheLocation.Directory}/{cacheLocation.FileName}_thumb.png";
+            if (this.fileSystem.File.Exists(checkPath))
+            {
+                return new SnapshotResult(false, checkPath);
+            }
 
             var outputPath = $"{cacheLocation.Directory}/{cacheLocation.FileName}_snap.png";
